Move bpm filter to its own file
