+++ packages/web/src/pages/search-page-v2/SearchFilters.tsx
@@ -1,4 +1,4 @@
-import { ReactElement, useCallback, useRef, useState } from 'react'
+import { ReactElement, useState } from 'react'
 
 import { GENRES, convertGenreLabelToValue } from '@audius/common/utils'
 import {
@@ -8,17 +8,14 @@ import {
   FilterButtonOptions,
   Popup,
   Paper,
-  TextInput,
   SegmentedControl,
   IconCaretDown,
   Divider,
-  Button,
-  Box,
-  useTheme
+  Box
 } from '@audius/harmony'
-import { css } from '@emotion/css'
 import { useSearchParams } from 'react-router-dom-v5-compat'
 
+import { BpmFilter } from './BpmFilter'
 import { Filter } from './types'
 import { MOODS } from './utils'
 
@@ -28,9 +25,6 @@ const messages = {
   mood: 'Mood',
   moodSearchPlaceholder: 'Search Mood',
   key: 'Key',
-  bpm: 'BPM',
-  minBpm: 'Min',
-  maxBpm: 'Max',
   isPremium: 'Premium',
   isVerified: 'Verified',
   hasDownloads: 'Downloads Available'
@@ -194,264 +188,6 @@ const KeyFilter = () => {
   )
 }
 
-type BpmTargetType = 'exact' | 'range5' | 'range10'
-const targetOptions: { label: string; value: BpmTargetType }[] = [
-  {
-    label: 'Exact',
-    value: 'exact'
-  },
-  {
-    label: '± 5',
-    value: 'range5'
-  },
-  {
-    label: '± 10',
-    value: 'range10'
-  }
-]
-const bpmOptions = [
-  {
-    label: 'Very Slow',
-    value: '1-60',
-    helperText: '<60 BPM'
-  },
-  {
-    label: 'Slow',
-    value: '60-90',
-    helperText: '60-90 BPM'
-  },
-  {
-    label: 'Medium',
-    value: '90-110',
-    helperText: '90-110 BPM'
-  },
-  {
-    label: 'Upbeat',
-    value: '110-140',
-    helperText: '110-140 BPM'
-  },
-  {
-    label: 'Fast',
-    value: '140-160',
-    helperText: '140-160 BPM'
-  },
-  {
-    label: 'Very Fast',
-    value: '160-999',
-    helperText: '160+ BPM'
-  }
-]
-
-const BpmFilter = () => {
-  const { color } = useTheme()
-  const [urlSearchParams] = useSearchParams()
-  const bpm = urlSearchParams.get('bpm')
-  const updateSearchParams = useUpdateSearchParams('bpm')
-  const [bpmFilterType, setBpmFilterType] = useState<'range' | 'target'>(
-    'range'
-  )
-  const [bpmTargetType, setBpmTargetType] = useState<BpmTargetType>('exact')
-
-  const minRef = useRef('')
-  const maxRef = useRef('')
-  const targetBpmRef = useRef('')
-  const targetTypeRef = useRef<BpmTargetType>('exact')
-
-  const handleBpmRangeChange = useCallback(
-    (handleChange: (value: string, label: string) => void) => {
-      let value = ''
-
-      if (minRef.current || maxRef.current) {
-        value = `${minRef.current || '1'}-${maxRef.current || '999'}`
-      }
-
-      handleChange(value, `${value} ${messages.bpm}`)
-    },
-    []
-  )
-
-  const handleBpmTargetChange = useCallback(
-    (handleChange: (value: string, label: string) => void) => {
-      let value = ''
-
-      if (targetBpmRef.current) {
-        if (targetTypeRef.current === 'exact') {
-          value = targetBpmRef.current
-        } else {
-          const mod = targetTypeRef.current === 'range5' ? 5 : 10
-          value = `${Math.max(
-            Number(targetBpmRef.current) - mod,
-            1
-          )}-${Math.min(Number(targetBpmRef.current) + mod, 999)}`
-        }
-      }
-
-      handleChange(value, `${value} ${messages.bpm}`)
-    },
-    []
-  )
-
-  const label = bpm ? `${bpm} ${messages.bpm}` : `${messages.bpm}`
-
-  return (
-    <FilterButton
-      value={bpm}
-      label={label}
-      onChange={updateSearchParams}
-      onReset={() => {
-        minRef.current = ''
-        maxRef.current = ''
-        targetBpmRef.current = ''
-      }}
-      iconRight={IconCaretDown}
-    >
-      {({ handleChange, isOpen, setIsOpen, anchorRef }) => (
-        <Popup
-          anchorRef={anchorRef}
-          isVisible={isOpen}
-          onClose={() => setIsOpen(false)}
-          anchorOrigin={{ vertical: 'bottom', horizontal: 'left' }}
-          transformOrigin={{ vertical: 'top', horizontal: 'left' }}
-        >
-          <Paper w={242} mt='s' border='strong' shadow='far'>
-            <Flex
-              w='100%'
-              pv='s'
-              gap='s'
-              direction='column'
-              alignItems='flex-start'
-              role='listbox'
-            >
-              <Flex
-                direction='column'
-                w='100%'
-                ph='s'
-                // NOTE: Adds a little flexibility so the user doesn't close the popup by accident
-                onClick={(e) => e.stopPropagation()}
-              >
-                <SegmentedControl
-                  options={[
-                    { key: 'range', text: 'Range' },
-                    { key: 'target', text: 'Target' }
-                  ]}
-                  selected={bpmFilterType}
-                  onSelectOption={(type) => {
-                    if (type === 'target') {
-                      minRef.current = ''
-                      maxRef.current = ''
-                    } else {
-                      targetBpmRef.current = ''
-                    }
-                    setBpmFilterType(type)
-                  }}
-                />
-              </Flex>
-              <Divider css={{ width: '100%' }} />
-              {bpmFilterType === 'range' ? (
-                <>
-                  <Flex direction='column' w='100%' ph='s'>
-                    <FilterButtonOptions
-                      options={bpmOptions}
-                      onChange={(option) => {
-                        handleChange(
-                          option.value,
-                          option.helperText ?? option.value
-                        )
-                      }}
-                    />
-                  </Flex>
-                  <Flex
-                    ph='s'
-                    gap='xs'
-                    w={240}
-                    alignItems='center'
-                    // NOTE: Adds a little flexibility so the user doesn't close the popup by accident
-                    onClick={(e) => e.stopPropagation()}
-                  >
-                    <TextInput
-                      label={messages.minBpm}
-                      placeholder={messages.minBpm}
-                      hideLabel
-                      onChange={(e) => {
-                        minRef.current = e.target.value
-                        handleBpmRangeChange(handleChange)
-                      }}
-                      inputRootClassName={css({
-                        height: '48px !important'
-                      })}
-                    />
-                    -
-                    <TextInput
-                      label={messages.maxBpm}
-                      placeholder={messages.maxBpm}
-                      hideLabel
-                      onChange={(e) => {
-                        maxRef.current = e.target.value
-                        handleBpmRangeChange(handleChange)
-                      }}
-                      inputRootClassName={css({
-                        height: '48px !important'
-                      })}
-                    />
-                  </Flex>
-                </>
-              ) : (
-                <Flex
-                  direction='column'
-                  ph='s'
-                  gap='s'
-                  // NOTE: Adds a little flexibility so the user doesn't close the popup by accident
-                  onClick={(e) => e.stopPropagation()}
-                >
-                  <TextInput
-                    label={messages.bpm}
-                    placeholder={messages.bpm}
-                    hideLabel
-                    onChange={(e) => {
-                      targetBpmRef.current = e.target.value
-                      handleBpmTargetChange(handleChange)
-                    }}
-                    inputRootClassName={css({
-                      height: '48px !important'
-                    })}
-                  />
-                  <Flex justifyContent='center' alignItems='center' gap='xs'>
-                    {targetOptions.map((option) => (
-                      <Button
-                        key={`targetOption_${option.value}`}
-                        size='small'
-                        variant={
-                          bpmTargetType === option.value
-                            ? 'primary'
-                            : 'tertiary'
-                        }
-                        // @ts-ignore
-                        hexColor={
-                          bpmTargetType === option.value
-                            ? color.secondary.secondary
-                            : undefined
-                        }
-                        fullWidth
-                        onClick={() => {
-                          setBpmTargetType(option.value)
-                          targetTypeRef.current = option.value
-                          handleBpmTargetChange(handleChange)
-                        }}
-                      >
-                        {option.label}
-                      </Button>
-                    ))}
-                  </Flex>
-                </Flex>
-              )}
-            </Flex>
-          </Paper>
-        </Popup>
-      )}
-    </FilterButton>
-  )
-}
-
 const IsPremiumFilter = () => {
   const [urlSearchParams] = useSearchParams()
   const isPremium = urlSearchParams.get('isPremium')
@@ -502,7 +238,7 @@ export const filters: Record<Filter, () => ReactElement> = {
   genre: GenreFilter,
   mood: MoodFilter,
   key: KeyFilter,
-  bpm: BpmFilter,
+  bpm: () => <BpmFilter useUpdateSearchParams={useUpdateSearchParams} />,
   isPremium: IsPremiumFilter,
   hasDownloads: HasDownloadsFilter,
   isVerified: IsVerifiedFilter
